agent/requirements.txt
google-adk
fastapi
uvicorn[standard]
pydantic
gunicorn
deprecated
psycopg2
pgvector